refactor to support different api backends
