+++ diagram-editor/Cargo.toml
@@ -44,8 +44,7 @@ serde = "1.0.219"
 serde_json = "1.0.140"
 tar = { version = "0.4.44", optional = true }
 thiserror = "2.0.16"
-tokio = { version = "1.45.1", features = ["macros", "time"] }
-tower-http = "0.6.6"
+tokio = { version = "1.45.1", features = ["macros"] }
 tracing = "0.1.41"
 
 [build-dependencies]
