+++ diagram-editor/frontend/api-client/index.ts
@@ -1,2 +1,3 @@
 export * from './api-client';
+export * from './base-api-client';
 export * from './debug-session';
