+++ diagram-editor/frontend/api-client/wasm/stub.d.ts
@@ -1,0 +1,6 @@
+// biome-ignore-all lint: generated
+/* tslint:disable */
+/* eslint-disable */
+export function init_wasm(): void;
+export function post_run(request: any): Promise<any>;
+export function get_registry(): any;
